ohlcv aggregation fix

+++ src/qubx/connectors/ccxt/handlers/ohlc.py
@@ -53,7 +53,7 @@ class OhlcDataHandler(BaseDataTypeHandler):
             oh[6] if len(oh) > 6 else 0.0,  # volume_quote
             oh[8] if len(oh) > 8 else 0.0,  # bought_volume
             oh[9] if len(oh) > 9 else 0.0,  # bought_volume_quote
-            float(oh[7]) if len(oh) > 7 else 0.0,  # trade_count
+            int(oh[7]) if len(oh) > 7 else 0,  # trade_count
         )
 
     def prepare_subscription(
